requirements.txt
colorzero==2.0
gpiozero==2.0.1
python-dotenv==1.0.1
